requirements.txt
sentence-transformers
faiss-cpu
transformers
torch
pandas
numpy
tqdm
python-dotenv
openai
langchain
langchain-community
chromadb
fastapi
uvicorn
langdetect
